add updates to support pkb_scheduler features

+++ perfkitbenchmarker/windows_virtual_machine.py
@@ -43,6 +43,10 @@ flags.DEFINE_bool(
     'set_cpu_priority_high', False,
     'Allows executables to be set to High (up from Normal) CPU priority '
     'through the SetProcessPriorityToHigh function.')
+flags.DEFINE_bool(
+    'skip_package_cleanup', False,
+    'skips cleanup')
+
 
 SMB_PORT = 445
 WINRM_PORT = 5986
@@ -449,7 +453,7 @@ class BaseWindowsMixin(virtual_machine.BaseOsMixin):
     self.home_dir = stdout.strip()
     stdout, _ = self.RemoteCommand('echo $env:SystemDrive')
     self.system_drive = stdout.strip()
-    self.RemoteCommand('mkdir %s' % self.temp_dir)
+    self.RemoteCommand('mkdir %s' % self.temp_dir, ignore_failure=True)
     self.DisableGuestFirewall()
 
   def _Reboot(self):
@@ -491,8 +495,11 @@ class BaseWindowsMixin(virtual_machine.BaseOsMixin):
     Deletes the Perfkit Benchmarker temp directory on the VM
     and uninstalls all PerfKit packages.
     """
+    if FLAGS.skip_package_cleanup:
+      return
     for package_name in self._installed_packages:
       self.Uninstall(package_name)
+
     self.RemoteCommand('rm -recurse -force %s' % self.temp_dir)
     self.EnableGuestFirewall()
